Setup jest for i18n, vuetify and filters
