+++ frontend/jest.config.js
@@ -17,4 +17,5 @@ module.exports = {
     "!**/dist/**",
   ],
   testMatch: ["**/*.spec.{ts,js}", "!**/node_modules/**"],
+  setupFiles: ["<rootDir>/src/setup-jest.ts"],
 };
